+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @redocly/openapi-core
 
+## 2.33.2
+
+## 2.33.1
+
+### Patch Changes
+
+- Improved AsyncAPI 3 types.
+- Updated js-yaml from `4.1.1` to `4.2.0`.
+
+## 2.33.0
+
+### Minor Changes
+
+- Added the `--component-names-strategy` option to the `bundle` command.
+  This option allows a choice of how inline Schema components are named: `basename` (default) or `title` (from each schema's `title` field).
+
 ## 2.32.2
 
 ## 2.32.1